python/modules/game_object.py
from typing import Any, Callable, Dict

class GameObject:
    def __init__(self, properties: Dict[str, Any], behaviors: Dict[str, Callable], parameters: Dict[str, Any]):
        self.properties = properties
        self.behaviors = behaviors
        self.parameters = parameters

    def update(self, dt: float):
        for behavior in self.behaviors.values():
            behavior(dt)

    def render(self):
        # This method is highly dependent on the specific game object and how it should be rendered.
        # Implement this method in a subclass based on the specific game object's rendering requirements.
        pass

    def add_behavior(self, name: str, behavior: Callable):
        self.behaviors[name] = behavior

    def remove_behavior(self, name: str):
        if name in self.behaviors:
            del self.behaviors[name]

    def set_property(self, name: str, value: Any):
        self.properties[name] = value

    def get_property(self, name: str) -> Any:
        return self.properties.get(name)

    def set_parameter(self, name: str, value: Any):
        self.parameters[name] = value

    def get_parameter(self, name: str) -> Any:
        return self.parameters.get(name)